orders/src/requirements.txt
aws-lambda-powertools
mypy-boto3-events
boto3-stubs
boto3-stubs[events]
